correct ZM_CONFIG_ with pkg_postinst ewarns
This is a breaking change.

Users need to complete the migration steps defined in pkg_postinst. ZM
was installing files in system etc paths vs its own applicaion dir.
This corrects that problem but does not attempt to migrate existing
configs created by the local sysadmin.

New installs are not impacted.
Thank you